Exception gandling

+++ src/main/java/com/devdioge/springcommerce/service/ProductService.java
@@ -3,11 +3,16 @@ package com.devdioge.springcommerce.service;
 import com.devdioge.springcommerce.dto.ProductDTO;
 import com.devdioge.springcommerce.entity.Product;
 import com.devdioge.springcommerce.repository.ProductRepository;
+import com.devdioge.springcommerce.service.exception.DatabaseException;
+import com.devdioge.springcommerce.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -22,7 +27,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findBytId(Long id) {
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Produto Nao Encontrado"));
         return modelMapper.map(product, ProductDTO.class);
     }
 
@@ -41,17 +47,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product product = productRepository.getReferenceById(id);
-        product.setName(dto.getName());
-        product.setDescription(dto.getDescription());
-        product.setPrice(dto.getPrice());
-        product.setImgUrl(dto.getImgUrl());
-        productRepository.save(product);
-        return modelMapper.map(product, ProductDTO.class);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            product.setName(dto.getName());
+            product.setDescription(dto.getDescription());
+            product.setPrice(dto.getPrice());
+            product.setImgUrl(dto.getImgUrl());
+            productRepository.save(product);
+            return modelMapper.map(product, ProductDTO.class);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Produto Nao Encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-       productRepository.deleteById(id);
+       if(!productRepository.existsById(id)){
+           throw new ResourceNotFoundException("Produto Nao Encontrado");
+       }
+       try {
+           productRepository.deleteById(id);
+       }
+       catch (DataIntegrityViolationException e) {
+           throw new DatabaseException("Falha de integridade referencial");
+       }
     }
 }
